Rename Visibility class, add kdoc
