+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/viewmodel/factory/MessageListViewModelFactory.kt
@@ -29,7 +29,7 @@ public class MessageListViewModelFactory @JvmOverloads constructor(
             MessageListViewModel(
                 cid,
                 messageId,
-                deletedMessageDisplayCondition = ChatUI.deletedMessageAppearance
+                deletedMessageVisibility = ChatUI.deletedMessageVisibility
             )
         },
     )
